requirements.txt
flask
flask-cors
torch
torchvision
pillow
numpy
timm
tqdm
scikit-learn
pyrebase
opencv-python
matplotlib
firebase-admin
pyrebase4
python-dotenv